requirements.txt
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-multipart>=0.0.5 